% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barChart.R
\name{amBarChart}
\alias{amBarChart}
\title{Bar chart}
\usage{
amBarChart(inputId, width = "100\%", height = "400px", data, category,
  value, valueNames = value, minValue, maxValue, valueFormatter = "#.",
  draggable = rep(FALSE, length(value)), chartTitle = NULL,
  columnStyle = list(fill = rep(list(NULL), length(value)), stroke =
  NULL, cornerRadius = NULL), backgroundColor = NULL, columnWidth = 80,
  xAxis = list(title = list(text = category, fontSize = 20, color =
  NULL), labels = list(color = NULL, fontSize = 18, rotation = 0)),
  yAxis = list(title = list(text = value, fontSize = 20, color = NULL),
  labels = list(color = NULL, fontSize = 18, rotation = 0)),
  gridLines = list(color = NULL, opacity = NULL, width = NULL),
  legend = length(value) > 1, caption = NULL, theme = NULL,
  style = "")
}
\arguments{
\item{inputId}{the id}

\item{width}{width in CSS units}

\item{height}{height in CSS units}

\item{data}{a dataframe}

\item{category}{name of the column of \code{data} to be used on the
category axis}

\item{value}{name(s) of the column(s) of \code{data} to be used on the
value axis}

\item{valueNames}{names of the values variables, to appear in the legend}

\item{minValue}{minimum value}

\item{maxValue}{maximum value}

\item{valueFormatter}{a number formatter; see
\url{https://www.amcharts.com/docs/v4/concepts/formatters/formatting-numbers/}}

\item{draggable}{logical vector, one entry for each value; whether the
corresponding colum is draggable}

\item{chartTitle}{chart title, \code{NULL}, character, or list of settings}

\item{columnStyle}{settings of the columns style given as a list}

\item{backgroundColor}{a HTML color for the chart background}

\item{columnWidth}{column width in percent}

\item{xAxis}{settings of the category axis given as a list, or just a string
for the axis title}

\item{yAxis}{settings of the value axis given as a list, or just a string
for the axis title}

\item{gridLines}{settings of the grid lines}

\item{legend}{logical, whether to display the legend}

\item{caption}{settings of the caption, or \code{NULL} for no caption}

\item{theme}{theme, \code{NULL} or one of \code{"dataviz"}, \code{"material"},
\code{"kelly"}, \code{"dark"}, \code{"moonrisekingdom"},
\code{"frozen"}, \code{"spiritedaway"}}

\item{style}{inline CSS for the container}
}
\description{
Constructs a bar chart.
}
\examples{
if(interactive()){

  # simple bar chart ####

  library(shiny)
  library(shinyAmBarCharts)

  dat <- data.frame(
    country = c("USA", "China", "Japan", "Germany", "UK", "France"),
    visits = c(3025, 1882, 1809, 1322, 1122, 1114)
  )

  ui <- fluidPage(
    fluidRow(
      column(8,
             amBarChart(
               "mybarchart", data = dat, height = "400px",
               category = "country", value = "visits",
               draggable = TRUE,
               chartTitle =
                list(text = "Visits per country", fontSize = 22, color = "orangered"),
               xAxis = list(title = list(text = "Country", color = "maroon")),
               yAxis = list(title = list(text = "Visits", color = "maroon")),
               minValue = 0, maxValue = 4000,
               valueFormatter = "#.",
               columnWidth = 90,
               caption = list(text = "Year 2018", color = "red"),
               theme = "material")),
      column(4,
             tags$label("Data:"),
             verbatimTextOutput("data"),
             br(),
             tags$label("Change:"),
             verbatimTextOutput("change"))
    )

  )

  server <- function(input, output){

    output[["data"]] <- renderPrint({
      input[["mybarchart"]]
    })

    output[["change"]] <- renderPrint({ input[["mybarchart_change"]] })

  }

  shinyApp(ui, server)

}

if(interactive()){

  # grouped bar chart ####

  library(shiny)
  library(shinyAmBarCharts)

  set.seed(666)
  dat <- data.frame(
    country = c("USA", "China", "Japan", "Germany", "UK", "France"),
    visits = c(3025, 1882, 1809, 1322, 1122, 1114),
    income = rpois(6, 25),
    expenses = rpois(6, 20)
  )

  ui <- fluidPage(

    fluidRow(
      column(8,
             amBarChart(
               "mygroupedbarchart", data = dat, height = "400px",
               category = "country", value = c("income", "expenses"),
               valueNames = c("Income", "Expenses"),
               draggable = c(TRUE,FALSE),
               backgroundColor = "#30303d",
               columnStyle = list(fill = c("darkmagenta", "darkred"),
                                  stroke = "#cccccc"),
               chartTitle = list(text = "Income and expenses per country"),
               xAxis = list(title = list(text = "Country")),
               yAxis = list(title = list(text = "Income and expenses")),
               minValue = 0, maxValue = 40,
               valueFormatter = "#.#",
               columnWidth = 90,
               caption = list(text = "Year 2018"),
               theme = "dark")
      ),
      column(4,
             tags$label("Data:"),
             verbatimTextOutput("data"),
             br(),
             tags$label("Change:"),
             verbatimTextOutput("change"))
    )

  )

  server <- function(input, output){

    output[["data"]] <- renderPrint({
      input[["mygroupedbarchart"]]
    })

    output[["change"]] <- renderPrint({ input[["mygroupedbarchart_change"]] })

  }

  shinyApp(ui, server)

}
}
